migrate Rpc code to RpcPublisher, delete Rpc
